+++ build.ps1
@@ -1,4 +1,4 @@
-$version = "0.5.8"
+$version = "0.6.0"
 
 $src = "$PSScriptRoot\src"
 $doc = "$PSScriptRoot\doc"