+++ src/Sandbox/SourceControlProvider.cs
@@ -1,22 +1,23 @@
-﻿using AstralKeks.PowerShell.Common.Provider;
+﻿//using AstralKeks.PowerShell.Common.Provider;
 using System.Management.Automation;
 using System.Management.Automation.Provider;
 using System.Collections.Generic;
 
 namespace AstralKeks.Workbench.Sandbox
 {
-    [CmdletProvider("SourceControl", ProviderCapabilities.None)]
-    [OutputType(typeof(string), ProviderCmdlet = ProviderCmdlet.ClearItem)]
-    public class SoureControlProvider : CmdletProviderBase
-    {
-        protected override IEnumerable<PSDriveInfo> OnInit()
-        {
-            yield return new PSDriveInfo("Src", ProviderInfo, "src:\\", "Source Control Drive", Credential);
-        }
+    //[CmdletProvider("SourceControl", ProviderCapabilities.None)]
+    //[OutputType(new[] { typeof(string) }, ProviderCmdlet = ProviderCmdlet.ClearItem)]
+    //public class SoureControlProvider : CmdletProviderBase
+    //{
+    //    protected override IEnumerable<PSDriveInfo> OnInit()
+    //    {
+    //        yield break;
+    //        //yield return new PSDriveInfo("Src", ProviderInfo, "src:\\", "Source Control Drive", Credential);
+    //    }
 
-        protected override CmdletProviderItem OnItem(string path)
-        {
-            return new SourceControlItem(path);
-        }
-    }
+    //    protected override CmdletProviderItem OnItem(string path)
+    //    {
+    //        return new SourceControlItem(path);
+    //    }
+    //}
 }