+++ src/Sandbox/TestCommand.cs
@@ -4,7 +4,7 @@ using System.Management.Automation;
 namespace AstralKeks.Workbench.Sandbox
 {
     [Cmdlet(VerbsDiagnostic.Test, "Command")]
-    [OutputType(typeof(string))]
+    [OutputType(new[] { typeof(string) })]
     public class TestCommand : Cmdlet
     {
         protected override void ProcessRecord()